removed TUM-ROSBAG.h, modified Makefile to remove TUM-ROSBAG targets and accept 'use_rosbag' instead.
